+++ backend/raonzena/src/main/java/com/ssafy/raonzena/api/service/UserServieImpl.java
@@ -124,7 +124,7 @@ public class UserServieImpl implements UserService{
     }
 
     @Override
-    public User selectUser(int userNo) {
+    public User selectUser(long userNo) {
         // userNo로 유저 정보 조회
         return userRepositorySupport.selectUser(userNo);
     }
